Fix for 32-bit host systems (bug #2267).
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@8989 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/src/stm8/gen.c
@@ -279,9 +279,9 @@ aopOnStackNotExt (const asmop *aop, int offset, int size)
 /* aopIsLitVal - asmop from offset is val                          */
 /*-----------------------------------------------------------------*/
 static bool
-aopIsLitVal (const asmop *aop, int offset, int size, unsigned long val)
+aopIsLitVal (const asmop *aop, int offset, int size, unsigned long long int val)
 {
-  wassert (size <= sizeof (unsigned long)); // Make sure we are not testing outside of argument val.
+  wassert (size <= sizeof (unsigned long long int)); // Make sure we are not testing outside of argument val.
 
   for(; size; size--, offset++)
     {
